so many changes dude!!!

+++ database/migrations/2020_07_06_181247_create_subcategories_table.php
@@ -16,8 +16,7 @@ class CreateSubcategoriesTable extends Migration
         Schema::create('subcategories', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedBigInteger('category_id');
-            // $table->foreign('category_id')->references('id')->on('categories');
+            $table->integer('category_id');
             $table->timestamps();
         });
     }
